+++ tests/unit/test_token_mint_transaction.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest.mock import MagicMock
 from hiero_sdk_python.tokens.token_mint_transaction import TokenMintTransaction
-from hiero_sdk_python.hapi.services import basic_types_pb2, timestamp_pb2, transaction_pb2, transaction_body_pb2
+from hiero_sdk_python.hapi.services import basic_types_pb2, timestamp_pb2, transaction_pb2, transaction_pb2
 from hiero_sdk_python.transaction.transaction_id import TransactionId
 from cryptography.hazmat.primitives import serialization
 from hiero_sdk_python.response_code import ResponseCode