+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 - ContractExecuteTransaction class
 - setMessageAndPay() function in StatefulContract
 - AccountDeleteTransaction Class
+- generate_proto.py
+- Bumped Hedera proto version from v0.57.3 to v0.64.3
 
 ### Changed
 - bump solo version to `v0.12`
@@ -24,6 +26,16 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 ### Removed
 - Removed the old `/documentation` folder.
 - Rebase command in README_upstream changed to just -S
+- generate_proto.sh
+
+### Breaking API changes
+- We have some changed imports and returns to maintain compatability in the proto bump 
+
+transaction_body_pb2.TransactionBody -> transaction_pb2.TransactionBody
+
+contract_call_local_pb2.ContractFunctionResult -> contract_types_pb2.ContractFunctionResult
+
+contract_call_local_pb2.ContractLoginfo -> contract_types_pb2.ContractLoginfo
 
 ## [0.1.4] - 2025-08-19
 ### Added